+++ src/events.ts
@@ -104,6 +104,7 @@ export interface BeaconEventType {
     p2pPeerInfo: P2PPairingRequest
     postmessagePeerInfo: PostMessagePairingRequest
     preferredNetwork: NetworkType
+    abortedHandler?(): void
   }
   [BeaconEvent.PAIR_SUCCESS]: ExtendedPostMessagePairingResponse | ExtendedP2PPairingResponse
   [BeaconEvent.P2P_CHANNEL_CONNECT_SUCCESS]: P2PPairingRequest
@@ -224,7 +225,11 @@ const showQrAlert = async (
   const alertConfig: AlertConfig = {
     title: 'Choose your preferred wallet',
     body: `<p></p>`,
-    pairingPayload: { p2pSyncCode: base58encoded, postmessageSyncCode: base58encoded, preferredNetwork: NetworkType.MAINNET }
+    pairingPayload: {
+      p2pSyncCode: base58encoded,
+      postmessageSyncCode: base58encoded,
+      preferredNetwork: NetworkType.MAINNET
+    }
   }
   await openAlert(alertConfig)
 }
@@ -245,7 +250,9 @@ const showPairAlert = async (data: BeaconEventType[BeaconEvent.PAIR_INIT]): Prom
       p2pSyncCode: p2pBase58encoded,
       postmessageSyncCode: postmessageBase58encoded,
       preferredNetwork: data.preferredNetwork
-    }
+    },
+    // eslint-disable-next-line @typescript-eslint/unbound-method
+    closeButtonCallback: data.abortedHandler
   }
   await openAlert(alertConfig)
 }