+++ src/types/beacon/NetworkType.ts
@@ -1,6 +1,6 @@
 export enum NetworkType {
   MAINNET = 'mainnet',
-  CARTHAGENET = 'carthagenet',
   DELPHINET = 'delphinet',
+  EDONET = 'edonet',
   CUSTOM = 'custom'
 }